Start using xarray/pandas (WIP)
From this commit, DSS files are being removed from the process. Instead,
xarray and pandas are being introduced. The work is in progress, so
results are not complete, for example, BBID values are not in the final
output, and they are slightly different from those of the previous
commit though program will run till the end. The changes are limited to
one of the modes, for DSM2 now. The post-processing part takes just a
few seconds now. The final outputs are in NetCDF files.