Add a Python version of DCD_kernel (WIP)
The DCD in Fortran (DCD_kernel.f) is reimplemented using Python.
Note that it is a WIP, and it does not implement a few routines in the
original Fortran codes, e.g. 1977 applied water adjustment and runoff
adjustment in leach water. Thus, results out of the Python version will
be different from that from the Fortran version.