Fix a bug and add runoff adjustment in leach water
In the previous commit, applied and drained leach water are not divided
by the number of days in month accidentally. This commit fixes it.

Routines to use runoff to reduce applied and drained leach water are now
implemented as it is don in the Fortran DCD. Some of logics in these
calculations are slightly modified to make them consistent.