Reimplement for_DCD and connect to new DCD (WIP)
Routines in for_DCD.py are reimplemented. Note that the reimplementation
is not complete, missing the so-called `extension` management. They need
to be addressed. The new implementation will read DETAW outputs in a
NetCDF file. This is connected to subsequent steps, formerly dcd_kernel,
and dcd_postprocessing, that were reimplemented already. Note that the
extension, SCHISM/CALSIM, double-counting adjustment, BBID splitting,
etc. are not reimplemented yet. The final output is a couple of NetCDF
files.
Now inputs for DCD are provided through a YAML file.
$ python dcd.py inputs/detaw.yaml

Two input files, GW_RATES.TXT and WYTYPES, are rolled back for the full
years.

DETAW-49