Modified the GW rates read function to accept yearly rates by island
The user can specify yearly gwrates by island and the updated function
translates this to daily values used by the model. This is an improvement
over the previous version which required preprocessing the daily values outside
of the model run. Now only the yearly values need to be updated by island.

[Ticket: X]