Split inputs for BBID and double-counting
This commit achieves BBID splitting and Calsim double-counting by
splitting areas (islands) and their information, instead of handling
those in the codes. This approach makes BBID splitting and
double-counting management simple, but a pre-processing step to split
information is necessary. A new file `util.py` is added for
pre-processing, but it is not complete yet.
While this change is implemented, most of the DCD input files are
converted into CSV formats instead of the original formats that are
different from file to file.
One new set of inputs for DSM2 CD are added. The result from this input
set is comparable to the original DETAW-DCD v1.3, but there are some
noticeable differences that are being reviewed.

DETAW-60