Change way to read ET correction
The commit slightly modifies the function to read ET correction file
because a new column, extension, is added. The old way selects to
columns to read by excluding columns to not read, which does not allow
us to add columns for additional information that may not be used by the
model.