Split areas only in DCD
This commit is a large breaking code change, aiming a next version,
v0.9. The goals of the update are to be more consistent to the previous
DETAW-DCD and to fix issues found in the current version.
Note that input files for the previous version will not work from now
on.

Here are changes:
1. The model now use unsplit subareas (or islands) in the DETAW part.
   The subareas get split in the DCD part, similarly to the previous
   DETAW-DCD did.
2. The ordering of the subtraction of those split subareas changed
   following the DETAW-DCD way.
3. The input files and the formats of those changed for the new way.

DETAW-96