Compute and write to output excess precip over water surface
Excess precip over water surface was previously being zeroed out. This did not matter much in the Delta since water surface areas are smaller, but in Suisun Marsh this was a large enough since there are significant water surface areas present. This is no longer zeroed out, and is written to the output netcdf as excess_precip_ws.

[Ticket: DETAW-114]